requirements.txt
asgiref==3.2.3
autopep8==1.5
Django==3.0.3
django-debug-toolbar==2.2
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
flake8-isort==2.8.0
isort==4.3.21
mccabe==0.6.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
sqlparse==0.3.0
testfixtures==6.12.0
toml==0.10.0
Werkzeug==1.0.0
Pillow